VEX-119: implemented #12

+++ core/src/main/java/cz/cleevio/core/widget/OfferLocationItem.kt
@@ -8,6 +8,7 @@ import cz.cleevio.core.databinding.WidgetOfferLocationItemBinding
 import cz.cleevio.repository.model.offer.Location
 import lightbase.core.extensions.layoutInflater
 import timber.log.Timber
+import java.math.BigDecimal
 
 class OfferLocationItem @JvmOverloads constructor(
 	context: Context,
@@ -25,18 +26,18 @@ class OfferLocationItem @JvmOverloads constructor(
 		binding = WidgetOfferLocationItemBinding.inflate(layoutInflater, this)
 	}
 
-	private fun getFloat(view: TextInputEditText): Float {
+	private fun getBigDecimal(view: TextInputEditText): BigDecimal {
 		return try {
-			view.text.toString().toFloat()
+			BigDecimal(view.text.toString())
 		} catch (exception: NumberFormatException) {
 			Timber.e(exception)
-			0.0f
+			BigDecimal(0.0)
 		}
 	}
 
 	fun getLocation(): Location = Location(
-		latitude = getFloat(binding.locationItemLatitude),
-		longitude = getFloat(binding.locationItemLongitude),
-		radius = getFloat(binding.locationItemRadius)
+		latitude = getBigDecimal(binding.locationItemLatitude),
+		longitude = getBigDecimal(binding.locationItemLongitude),
+		radius = getBigDecimal(binding.locationItemRadius)
 	)
 }
