+++ repository/src/main/java/cz/cleevio/repository/model/offer/Location.kt
@@ -1,10 +1,11 @@
 package cz.cleevio.repository.model.offer
 
 import com.squareup.moshi.JsonClass
+import java.math.BigDecimal
 
 @JsonClass(generateAdapter = true)
 data class Location constructor(
-	val longitude: Float,
-	val latitude: Float,
-	val radius: Float
+	val longitude: BigDecimal,
+	val latitude: BigDecimal,
+	val radius: BigDecimal
 )
